Adds many small QOL improvements:
As soon as you load an image, the selected item in the overlay dropdown box is applied right away. This can speed things up if you have a lot of images to do one by one.

Quick overwrite save button to overwrite current image you are working on.

Additional Bluray 3D and Bluray 3D IMAX overlays.

"Reset" button is colored and stands out more. It also disposes of the image from memory so that the image can be deleted from file explorer while still having the app open. This can be helpful in some cases.

"Save" button disposes from memory the original image you opened, overwrites it with a new copy of the changed image and then re-opens it so you can continue to work on it.

"Save as" button disposes from memory the original image you opened and lets you save the new image elsewhere and re-opens that newly saved image instead so you can continue to work off that. This unlocks the old image so that it can be deleted in file explorer while still having the app open. This can be helpful in some cases depending on what user is trying to do.